@@ -1,5 +1,5 @@
 import 'package:flutter/material.dart';
-import 'package:page_view_indicator/page_view_indicator.dart';
+import 'package:page_view_indicator_ns/page_view_indicator.dart';
 
 class App extends StatelessWidget {
   static const length = 3;
@@ -45,48 +45,6 @@ class App extends StatelessWidget {
               size: 12.0,
               color: Colors.accents.elementAt((index + 3) * 3),
             ),
-          ),
-    );
-  }
-
-  PageViewIndicator _buildExample2() {
-    return PageViewIndicator(
-      pageIndexNotifier: pageIndexNotifier,
-      length: length,
-      normalBuilder: (animationController, index) => Circle(
-            size: 8.0,
-            color: Colors.black87,
-          ),
-      highlightedBuilder: (animationController, index) => ScaleTransition(
-            scale: CurvedAnimation(
-              parent: animationController,
-              curve: Curves.ease,
-            ),
-            child: Circle(
-              size: 8.0,
-              color: Colors.white,
-            ),
-          ),
-    );
-  }
-
-  PageViewIndicator _buildExample3() {
-    return PageViewIndicator(
-      pageIndexNotifier: pageIndexNotifier,
-      length: length,
-      normalBuilder: (animationController, index) => ScaleTransition(
-            scale: CurvedAnimation(
-              parent: animationController,
-              curve: Curves.ease,
-            ),
-            child: Icon(Icons.favorite, color: Colors.black87),
-          ),
-      highlightedBuilder: (animationController, index) => ScaleTransition(
-            scale: CurvedAnimation(
-              parent: animationController,
-              curve: Curves.ease,
-            ),
-            child: Icon(Icons.star, color: Colors.white),
           ),
     );
   }
